fix: Resolve all SonarAnalyzer errors for clean build
Fix all analyzer violations to achieve 0 errors, 0 warnings with
SonarAnalyzer.CSharp v10.17.0 and TreatWarningsAsErrors enabled.

Key fixes:
- S2325: Make methods static where appropriate
- S1066: Merge nested if statements
- S1481: Replace unused variables with discards
- S127: Refactor for loops to while loops in CLI arg parsing
- S3218: Rename properties that shadow System types
- S3878: Suppress where conflicting with S3220 (params array)
- S3904: Suppress for SDK projects (no assembly output)
- S3881: Implement proper IDisposable pattern
- S2292: Convert to auto-implemented properties
- IDE0059: Remove unnecessary variable assignments

Also updates test files to use static method calls after
refactoring instance methods to static.

+++ src/KeenEyes.Navigation.DotRecast/DotRecastProvider.cs
@@ -170,7 +170,7 @@ public sealed class DotRecastProvider : INavigationProvider
             startPt, endPt,
             pathList, pathList.Count,
             straightPath, out var straightPathCount, 256,
-            (int)DtStraightPathOptions.DT_STRAIGHTPATH_ALL_CROSSINGS);
+            DtStraightPathOptions.DT_STRAIGHTPATH_ALL_CROSSINGS);
 
         if (status.Failed() || straightPathCount == 0)
         {
